Adding Cart And Order

+++ db/migrate/20220720075936_add_column_to_product.rb
@@ -1,7 +1,7 @@
 class AddColumnToProduct < ActiveRecord::Migration[7.0]
   def change
     add_column :products, :title, :string
-    add_column :products, :price, :decimal
+    add_column :products, :price, :float
     add_column :products, :description, :text
   end
 end
